SDK/test-deployment/requirements.txt
# Test deployment requirements
guardflow-fastapi
uvicorn[standard]